requirements.txt
asgiref==3.8.1
Django==5.1.3
django-autoslug==1.9.9
psycopg-binary==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
